Extra output to controol 'disable feed-in' to grid
When the prices for the current hour are negative, the output will
output 1, otherwise 0. So the output can be connected to the
ESS Control node: Disable feed-in.

This can be overruled via either configuration (allow disable feed-in)
and via sending a control message on `msg.allow_disable_feedin` containing
the desired boolean value.