Cope with 48 hour forecast at midnight
If at midnight and there was (still) a 48 hour schedule to work with,
the system picked the target SOC at hour 0 instead of the one at
hour 24. This happened only until the next schedule update, so only
for a few minutes the system worked with the wrong target SOC, but
it still an annoying bug.