Add support for selecting UK Octopus contracts
Separate buy and sell contract for Octopus users, allowing to schedules to change on the half hour.
Adding tariff and region to the contract_buy and contract_sell parameters. Note that Octopus users will need to figure out the tariff and the region. The region is the last letter on the name of the contract.